pyproject.toml
[project]
name = "repo-onboarding-agent"
version = "0.1.0"
description = "A LangGraph-powered agent that explores GitHub repositories and produces structured onboarding guides"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "gitpython>=3.1",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "redis>=5.0",
    "sse-starlette>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]
finetune = [
    "mlx-lm>=0.18",
    "rouge-score>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
